requirements-dev.txt
asgiref==3.5.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.4
Django-Verify-Email==2.0.3
gunicorn==20.1.0
python-decouple==3.6
python-dotenv==0.20.0
sqlparse==0.4.2
static3==0.7.0
